backend/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
psycopg2-binary
pydantic
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
pandas
scikit-learn
joblib
python-multipart
python-dotenv
google-generativeai